+++ app/Http/Controllers/Backend/SubCategoryController.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\SubCategoryStoreRequest;
 use App\Http\Requests\Admin\SubCategoryUpdateRequest;
 use App\Models\Category;
+use App\Models\ChildCategory;
 use App\Models\SubCategory;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
@@ -92,6 +93,15 @@ class SubCategoryController extends Controller
      */
     public function destroy(SubCategory $subCategory)
     {
+        // Check if the category has any associated subcategories
+        $childCategory = ChildCategory::where('sub_category_id', $subCategory->id)->count();
+
+        if ($childCategory > 0){
+            return response([
+                'status' => 'error',
+                'message' => 'This items contain, sub items for delete this you have to delete the sub items first!'
+            ]);
+        }
         $subCategory->delete();
 
         return response(['status' => 'success', 'message' => 'Sub Category deleted successfully']);
